requirements.txt
requests==2.26.0
pandas==1.3.1
tldextract==3.2.0
PyYAML==5.4.1
BeautifulSoup4==4.10.0
lxml==4.6.3
openpyxl==3.0.9
